# api/openapi.yaml
openapi: 3.0.3
info:
  title: Simplus API
  version: 0.1.0
  description: Locale-neutral control-plane contract for Simplus.
servers:
  - url: /
paths:
...
  /api/v1/notification-channels:
    get:
      operationId: listNotificationChannels
      tags: [notifications]
      summary: List outbound notification channels without credentials
      responses:
        "200": { description: Channels, content: { application/json: { schema: { $ref: "#/components/schemas/NotificationChannelList" } } } }
    post:
      operationId: createNotificationChannel
      tags: [notifications]
      summary: Store an encrypted WeCom or Feishu webhook
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/NotificationChannelMutation" } } } }
      responses:
        "201": { description: Created channel, content: { application/json: { schema: { $ref: "#/components/schemas/NotificationChannel" } } } }
        "400": { description: Invalid channel, content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } } }
  /api/v1/notification-channel-bindings/feishu:
    get:
      operationId: getFeishuNotificationBinding
      tags: [notifications]
      summary: Read the current transient Feishu private-message binding attempt
      responses:
        "200": { description: Current binding state, headers: { Cache-Control: { schema: { type: string } } }, content: { application/json: { schema: { $ref: "#/components/schemas/FeishuNotificationBinding" } } } }
        "503": { description: Binding service unavailable, content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } } }
    post:
      operationId: startFeishuNotificationBinding
      tags: [notifications]
      summary: Start one create-only least-privilege Feishu application binding attempt
      responses:
        "201": { description: Waiting for administrator authorization, headers: { Cache-Control: { schema: { type: string } } }, content: { application/json: { schema: { $ref: "#/components/schemas/FeishuNotificationBinding" } } } }
        "409": { description: Another binding is active, content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } } }
        "502": { description: Feishu registration could not be started, content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } } }
    delete:
      operationId: cancelFeishuNotificationBinding
      tags: [notifications]
      summary: Cancel a waiting Feishu binding attempt
      responses:
        "200": { description: Current binding state, headers: { Cache-Control: { schema: { type: string } } }, content: { application/json: { schema: { $ref: "#/components/schemas/FeishuNotificationBinding" } } } }
        "409": { description: The attempt is already testing and cannot be cancelled, content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } } }
  /api/v1/notification-channels/{channelId}:
    parameters:
      - { name: channelId, in: path, required: true, schema: { type: string, pattern: "^channel_[A-Za-z0-9_-]{22}$" } }
    put:
      operationId: updateNotificationChannel
      tags: [notifications]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/NotificationChannelMutation" } } } }
      responses:
        "200": { description: Updated channel, content: { application/json: { schema: { $ref: "#/components/schemas/NotificationChannel" } } } }
        "404": { description: Channel not found, content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } } }
    delete:
      operationId: deleteNotificationChannel
      tags: [notifications]
      responses:
        "204": { description: Deleted }
        "404": { description: Channel not found, content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } } }
  /api/v1/notification-channels/{channelId}/test:
    parameters:
      - { name: channelId, in: path, required: true, schema: { type: string, pattern: "^channel_[A-Za-z0-9_-]{22}$" } }
    post:
      operationId: testNotificationChannel
      tags: [notifications]
      summary: Send one explicit test message
      responses:
        "200": { description: Updated delivery status, content: { application/json: { schema: { $ref: "#/components/schemas/NotificationChannel" } } } }
        "404": { description: Channel not found, content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } } }
        "502": { description: Provider rejected or could not receive the test, content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } } }
...
components:
...
  schemas:
...
    NotificationEventKind:
      type: string
      enum: [sms.received, sms.failed, call.incoming, call.missed, system.degraded, vowifi.connected, vowifi.disconnected, cellular.connected, cellular.disconnected]
    NotificationChannel:
      type: object
      additionalProperties: false
      required: [id, provider, deliveryMode, targetType, displayName, webhookHint, signingSecretConfigured, enabled, eventKinds, lastDeliveryAt, lastDeliveryStatus, lastErrorCode, pendingCount, failedCount]
      properties:
        pendingCount: { type: integer, format: int64, minimum: 0 }
        failedCount: { type: integer, format: int64, minimum: 0 }
        id: { type: string, pattern: "^channel_[A-Za-z0-9_-]{22}$" }
        provider: { type: string, enum: [wecom, feishu] }
        deliveryMode: { type: string, enum: [webhook, feishu_app] }
        targetType: { type: string, enum: [webhook, authorized_user] }
        displayName: { type: string, minLength: 1, maxLength: 80 }
        webhookHint: { type: string, enum: [qyapi.weixin.qq.com, open.feishu.cn] }
        signingSecretConfigured: { type: boolean }
        enabled: { type: boolean }
        eventKinds: { type: array, minItems: 1, maxItems: 9, uniqueItems: true, items: { $ref: "#/components/schemas/NotificationEventKind" } }
        lastDeliveryAt: { type: string, maxLength: 64 }
        lastDeliveryStatus: { type: string, enum: [never, success, failed] }
        lastErrorCode: { type: string, maxLength: 64 }
    NotificationChannelMutation:
      type: object
      additionalProperties: false
      required: [provider, displayName, webhookUrl, signingSecret, enabled, eventKinds]
      properties:
        provider: { type: string, enum: [wecom, feishu] }
        displayName: { type: string, minLength: 1, maxLength: 80 }
        webhookUrl: { type: string, maxLength: 4096, description: Required on create; empty on update preserves the credential }
        signingSecret: { type: string, maxLength: 512, description: Optional Feishu signature secret; empty preserves the existing secret }
        enabled: { type: boolean }
        eventKinds: { type: array, minItems: 1, maxItems: 9, uniqueItems: true, items: { $ref: "#/components/schemas/NotificationEventKind" } }
    NotificationChannelList:
      type: object
      additionalProperties: false
      required: [channels]
      properties:
        channels: { type: array, maxItems: 20, items: { $ref: "#/components/schemas/NotificationChannel" } }
    FeishuNotificationBinding:
      type: object
      additionalProperties: false
      required: [state, verificationUrl, expiresAt, channelId, errorCode]
      properties:
        state: { type: string, enum: [idle, waiting, testing, succeeded, failed, expired, cancelled] }
        verificationUrl: { type: string, maxLength: 4096, description: Present only while waiting; never persisted }
        expiresAt: { type: string, maxLength: 64 }
        channelId: { type: string, pattern: "^$|^channel_[A-Za-z0-9_-]{22}$" }
        errorCode: { type: string, maxLength: 64, pattern: "^$|^[A-Z0-9_]+$" }
...
    ApiError:
      type: object
      additionalProperties: false
      required: [code, retryable]
      properties:
        code:
          type: string
          pattern: "^[A-Z0-9_]+$"
        retryable:
          type: boolean
        reference:
          type: string